+++ w4/army/spell/Spell.cpp
@@ -1,6 +1,6 @@
 #include "Spell.h"
 
-Spell::Spell(MagicState* mState, const std::string& spellName, float cost, float points) : mState(mState), spellName(spellName), cost(cost), points(points) {};
+Spell::Spell(MagicState* mState, const std::string& spellName, double cost, double points) : mState(mState), spellName(spellName), cost(cost), points(points) {};
 
 Spell::~Spell() {};
 
@@ -8,10 +8,10 @@ const std::string& Spell::getSpellName() const {
     return this->spellName;
 };
 
-float Spell::getCost() const {
+double Spell::getCost() const {
     return this->cost;
 };
 
-float Spell::getPoints() const {
+double Spell::getPoints() const {
     return this->points;
 };
