+++ w4/army/state/BerserkerState.cpp
@@ -4,6 +4,6 @@ BerserkerState::BerserkerState() : HumanState(150, 1.5, 0.5) {};
 
 BerserkerState::~BerserkerState() {};
 
-void BerserkerState::takeMagicDamage(float magicDmg) {
+void BerserkerState::takeMagicDamage(double magicDmg) {
     std::cout << "Berserker cannot takeMagicDamage()!" << std::endl;
 };
