+++ w4/army/state/WerewolfState.cpp
@@ -1,10 +1,9 @@
 #include "WerewolfState.h"
 
-WerewolfState::WerewolfState(float hp) : State(hp) {
+WerewolfState::WerewolfState(double hp) : State(hp) {
     this->isVampire = false;
     this->isUndead = false;
     this->isWerewolf = true;
-    // this->isWolf = false;
 
     this->strength = 1.25;
     this->agility = 0.75;
