+++ w4/army/state/WerewolfState.h
@@ -6,7 +6,7 @@
 
 class WerewolfState : public State {
     public:
-        WerewolfState(float hp = 100);
+        WerewolfState(double hp = 100);
         virtual ~WerewolfState();
 };
